Change sensor restrictToColor from single color to 10-color bitset

+++ source/EngineGpuKernels/Entities.cuh
@@ -103,14 +103,14 @@ struct DetectStructure
 struct DetectFreeCell
 {
     float minDensity;
-    uint8_t restrictToColor;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint16_t restrictToColor;  // bitset: bit i set = color i allowed, 0x3FF = all colors
 };
 
 struct DetectCreature
 {
-    uint32_t minNumCells;     // 0 = no restriction
-    uint32_t maxNumCells;     // 0 = no restriction
-    uint8_t restrictToColor;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint32_t minNumCells;      // 0 = no restriction
+    uint32_t maxNumCells;      // 0 = no restriction
+    uint16_t restrictToColor;  // bitset: bit i set = color i allowed, 0x3FF = all colors
     LineageRestriction restrictToLineage;
 };
