+++ source/EngineGpuKernels/Genome.cuh
@@ -40,14 +40,14 @@ struct DetectStructureGenome
 struct DetectFreeCellGenome
 {
     float minDensity;
-    uint8_t restrictToColor;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint16_t restrictToColor;  // bitset: bit i set = color i allowed, 0x3FF = all colors
 };
 
 struct DetectCreatureGenome
 {
-    uint32_t minNumCells;     // 0 = no restriction
-    uint32_t maxNumCells;     // 0 = no restriction
-    uint8_t restrictToColor;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint32_t minNumCells;      // 0 = no restriction
+    uint32_t maxNumCells;      // 0 = no restriction
+    uint16_t restrictToColor;  // bitset: bit i set = color i allowed, 0x3FF = all colors
     LineageRestriction restrictToLineage;
 };
